pyproject.toml
[project]
name = "katalab"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google>=3.0.0",
    "google-genai>=1.55.0",
    "pydantic>=2.12.5",
    "pydantic-ai>=1.32.0",
    "python-dotenv>=1.2.1",
    "ruff>=0.14.9",
]
